requirements.txt
PySide6>=6.8
python-dotenv>=1.0
httpx>=0.27
python-socketio[asyncio_client]>=5.11
websocket-client>=1.8
sgp4>=2.23

# --- build-time only (not needed at runtime) ---
Pillow>=10.0
pyinstaller>=6.0
